UI: have the input passthrough containers
UI layout can be deep and having the output of inner shards bubble up is confusing and could lead to unexpected results.
It is also unlikely to be practical as widgets are already organized in a logical fashion.
Users can always introduce a variable if some inner output needs to be used elsewhere.