Refactor Handle and add support for disc hit raycasting
These changes add support for resolving raycasting to check on hits on handles that use discs rather than just boxes.

Currently uses virtual functions for handles to support different behaviour depending on whether handles are boxes or discs, but up for better suggestions for implementation.

Currently, resolveHover for DiscHandle does not appear to work properly and is being worked on. There appears to be an issue with the raycast/plane that the hits are being registered on, resulting in a very different collisionbox from the expected disc to be rendered