Add RotationGizmo shard and behaviour and change translation gizmo to updated handle
These changes add the shard for the RotationGizmo, and specific behaviour for the RotationGizmo. The TranslationGizmo has some changes to reflect the updated Handle which is now a virtual class.

For RotationGizmo, I'm not sure if there is a built-in rotation matrix in linalg, it seems like most of the functions in the library regarding rotation is meant for quarternions, including linalg::rotation_matrix. Currently, rotation works, some adjustments will have to be made regarding how much movement when dragging the gizmo handles translate to how much rotation, but there is still the issue to fix with raycasting on discs not being correct.

There may be a lot of irrelevant comments due to WIP.

Some portions of code were done with the support for rotated handles of the RotationGizmo, but whether the logic is sound remains unverified and to be checked only when a base working version of the RotationGizmo is available.