Add MetalViewContainer event for Apple devices in main window
This commit adds the ability to get the MetalViewContainer for Apple devices in the main window. This allows the graphics renderer to access the Metal framework API for rendering on Apple devices.

This is a simple hack for now, eventually we want a more structured way to access Shards data