feat(build): Enable native architecture support in macOS build
This commit updates the macOS build script to add "-DARCH=native" to the CMake command to enable native architecture support when building the software on macOS machines. Additionally, the commit updates the default architecture name from "core2" to "pentium4" in the cmake/Platform.cmake file when compiling for x86 on Windows.

# Conflicts:
#	shards/gfx/rust/naga
#	shards/gfx/rust/profiling
#	shards/gfx/rust/wgpu
#	shards/gfx/rust/wgpu-native