Add sub expressions
Add default shard before evaluating new pipeline blocks

This commit adds an automatically generated nil shard at the start of a new pipeline. This behavior accompanies many shards such as Http. If the pipeline already has a previous shard, the nil is ignored. A new function, `add_const_shard2`, has been added to accomplish the behavior. The update also includes a conditional that removes the nil shard if the pipeline's first block is a constant. The previous shard is set to `None` before evaluating a new pipeline. A test case has also been added to ensure the behavior of the nil shard.

add EvalExpr but we missed bools etc

add operators, bools and none