Fix weird behavior of std::optional<tf::Exector>
It works with unique_ptr, maybe some anomaly in windows implementation of optional<>

Copilot says: "Copilot thinks that the `tf::Executor` is not default constructible,
which is not true. It is default constructible, but it is not
copyable. So, we need to use `std::unique_ptr` instead of
`std::optional`." :)